requirements.txt
asgiref==3.5.2
Babel==2.10.1
Django==3.2.13
django-extra-views==0.13.0
django-haystack==3.2.1
django-oscar==3.1
django-phonenumber-field==5.2.0
django-tables2==2.3.4
django-treebeard==4.4
django-widget-tweaks==1.4.12
factory-boy==3.2.1
Faker==13.12.0
phonenumbers==8.12.48
Pillow==9.1.1
purl==1.6
pycountry==22.3.5
python-dateutil==2.8.2
python-memcached==1.59
pytz==2022.1
reportlab==3.6.9
six==1.16.0
sorl-thumbnail==12.8.0
sqlparse==0.4.2
tinycss2==1.1.1
webencodings==0.5.1
